+++ graphs.product.set.cpp
@@ -35,13 +35,13 @@ GraphsProductSet::GraphsProductSet( const GraphsProduct& gp )
 		graphs.insert( *it );
 }
 
-vector< unsigned int > GraphsProductSet::get_iVertices( ) const
+vector< short unsigned int > GraphsProductSet::get_iVertices( ) const
 {
-	vector< unsigned int > iVertices;
+	vector< short unsigned int > iVertices;
 	
 	for( set< Graph* >::const_iterator itProd( graphs.begin( ) ); itProd != graphs.end( ); ++itProd )
 	{
-		for( vector< unsigned int >::const_iterator itV( (*itProd)->iVertices.begin( ) ); itV != (*itProd)->iVertices.end( ); ++itV )
+		for( vector< short unsigned int >::const_iterator itV( (*itProd)->iVertices.begin( ) ); itV != (*itProd)->iVertices.end( ); ++itV )
 		{
 			iVertices.push_back( *itV );
 		}
@@ -56,9 +56,9 @@ bool GraphsProductSet::b_areVerticesSubsetOf( const GraphsProductSet& gp ) const
 	/* NOTE:
 		* We could sort de vectors and use the standard function includes but this is slower (tested with 17-vinb85).
 	*/
-	vector< unsigned int > iVerticesSmall( get_iVertices( ) ), iVerticesBig( gp.get_iVertices( ) );
+	vector< short unsigned int > iVerticesSmall( get_iVertices( ) ), iVerticesBig( gp.get_iVertices( ) );
 	
-	for( vector< unsigned int >::const_iterator it( iVerticesSmall.begin( ) ); it != iVerticesSmall.end( ); ++it )
+	for( vector< short unsigned int >::const_iterator it( iVerticesSmall.begin( ) ); it != iVerticesSmall.end( ); ++it )
 	{
 		if( find( iVerticesBig.begin( ), iVerticesBig.end( ), *it ) == iVerticesBig.end( ) )
 			return false;
